refactor: extract mod_settings into module
